fix(simulation): 消除仿真层 f32 泄漏，改为万分比整数运算
- 所有概率/比率配置从 f32 改为 u32 万分比（combat.ron/units.ron/cities.ron）
- 新增 gen_probability_permyriad() 替代 gen_probability()
- 骑兵闪避、吸血、减速倍率、城市伤害/治疗/升级等全部改为整数乘除
- Arrow.pierce_chance 从 f32 改为 u32
- AI 攻击决策比较改为整数（10:13 比例）
- nearest-enemy 扫描 HashMap 改为 BTreeMap
- 全部 83 个测试通过

+++ crates/simulation/src/combat/config.rs
@@ -8,7 +8,7 @@ use serde::Deserialize;
 pub struct ShieldConfig {
     pub speed_penalty: u32,
     pub attack_speed_penalty: u32,
-    pub passive_block_chance: f32,
+    pub passive_block_chance: u32,
     pub frontal_angle_deg: u32,
     pub initial_hp: u32,
     pub drop_survive_ticks: u32,
@@ -18,25 +18,25 @@ pub struct ShieldConfig {
 /// Cavalry dodge mechanics.
 #[derive(Clone, Debug, Deserialize)]
 pub struct CavalryConfig {
-    pub dodge_max_chance: f32,
-    pub dodge_decay_rate: f32,
+    pub dodge_max_chance: u32,
+    pub dodge_decay_rate: u32,
 }
 
 /// Archer multi-shot mechanics.
 #[derive(Clone, Debug, Deserialize)]
 pub struct ArcherMultiShotConfig {
-    pub base_chance: f32,
-    pub per_level_bonus: f32,
-    pub min_chance: f32,
-    pub max_chance: f32,
+    pub base_chance: u32,
+    pub per_level_bonus: u32,
+    pub min_chance: u32,
+    pub max_chance: u32,
 }
 
 /// Slow debuff applied by archer arrows.
 #[derive(Clone, Debug, Deserialize)]
 pub struct SlowDebuffConfig {
-    pub base_amount: f32,
-    pub stack_mult: f32,
-    pub max_reduction: f32,
+    pub base_amount: u32,
+    pub stack_mult: u32,
+    pub max_reduction: u32,
     pub duration_ticks: u32,
     pub max_stacks: u32,
 }
@@ -49,7 +49,7 @@ pub struct LevelUpConfig {
     pub hp_gain: u32,
     pub attack_gain: u32,
     pub lifesteal_unlock_level: u32,
-    pub lifesteal_rate: f32,
+    pub lifesteal_rate: u32,
 }
 
 /// Fearless buff (triggered on cavalry dodge).
@@ -57,7 +57,7 @@ pub struct LevelUpConfig {
 pub struct FearlessConfig {
     pub duration_ticks: u32,
     pub attack_bonus: u32,
-    pub lifesteal_bonus: f32,
+    pub lifesteal_bonus: u32,
 }
 
 /// Post-tick overlap resolution — guarantees no soldiers overlap after each tick.
@@ -82,8 +82,8 @@ pub struct AttackWindupConfig {
 /// Top-level combat configuration.
 #[derive(Clone, Debug, Deserialize, Resource)]
 pub struct CombatGlobalConfig {
-    pub city_damage_per_soldier_ratio: f32,
-    pub arrow_building_damage_ratio: f32,
+    pub city_damage_per_soldier_ratio: u32,
+    pub arrow_building_damage_ratio: u32,
     pub facing: FacingConfig,
     pub attack_windup: AttackWindupConfig,
     pub shield: ShieldConfig,
